Adding extention verification and eicar execution

+++ Diagnostics/AVTester/Test-ExchAVExclusions.ps1
@@ -31,7 +31,7 @@ Log file:
 $env:LOCALAPPDATA\ExchAvExclusions.log
 
 List of Scanned Folders:
-$env:LOCALAPPDATA\BadFolders.txt
+$env:LOCALAPPDATA\BadExclusions.txt
 
 .EXAMPLE
 .\Test-ExchAVExclusions.ps1
@@ -105,7 +105,7 @@ if (-not($exchangeShell.ShellLoaded)) {
 }
 
 # Create the Array List
-$BaseFolders = Get-ExchAVExclusions -ExchangePath $ExchangePath
+$BaseFolders = Get-ExchAVExclusionsPaths -ExchangePath $ExchangePath -MsiProductMinor ([byte]$serverExchangeInstallDirectory.MsiProductMinor)
 
 if ( $BaseFolders.count -eq 0 ) {
     Write-Warning "We do not detect folders to analyze"
@@ -126,28 +126,33 @@ foreach ($path in $BaseFolders) {
         if ($Recurse) {
 
             # Add the root folder
-            $FolderList.Add($path)
+            $FolderList.Add($path.tolower())
 
             # Get the Folder and all subFolders and just return the fullname value as a string
-            Get-ChildItem $path -Recurse -Directory -ErrorAction SilentlyContinue | Select-Object -ExpandProperty FullName | ForEach-Object { $FolderList.Add($_) }
+            Get-ChildItem $path -Recurse -Directory -ErrorAction SilentlyContinue | Select-Object -ExpandProperty FullName | ForEach-Object { $FolderList.Add($_.tolower()) }
         }
         # Just Add the root folder
-        else { $FolderList.Add($path) }
+        else { $FolderList.Add($path.tolower()) }
     } catch { Write-SimpleLogfile -string ("[ERROR] - Failed to resolve folder " + $path) -Name $LogFile }
 }
+
+# Remove any Duplicates
+$FolderList = $FolderList | Select-Object -Unique
 
 Write-SimpleLogfile -String "Creating EICAR Files" -name $LogFile -OutHost
 
 # Create the EICAR file in each path
-$EicarFileName = "eicar.bat"
+$EicarFileName = "eicar"
+$EicarFileExt = "com"
+$EicarFullFileName = "$EicarFileName.$EicarFileExt"
+
+#Base64 of Eicar string
+[string] $EncodedEicar = 'WDVPIVAlQEFQWzRcUFpYNTQoUF4pN0NDKTd9JEVJQ0FSLVNUQU5EQVJELUFOVElWSVJVUy1URVNULUZJTEUhJEgrSCo='
 
 foreach ($Folder in $FolderList) {
 
-    [string] $FilePath = (Join-Path $Folder $EicarFileName)
-    Write-SimpleLogfile -String ("Creating EICAR file " + $FilePath) -name $LogFile
-
-    #Base64 of Eicar string
-    [string] $EncodedEicar = 'WDVPIVAlQEFQWzRcUFpYNTQoUF4pN0NDKTd9JEVJQ0FSLVNUQU5EQVJELUFOVElWSVJVUy1URVNULUZJTEUhJEgrSCo='
+    [string] $FilePath = (Join-Path $Folder $EicarFullFileName)
+    Write-SimpleLogfile -String ("Creating $EicarFullFileName file " + $FilePath) -name $LogFile
 
     if (!(Test-Path -Path $FilePath)) {
 
@@ -159,25 +164,73 @@ foreach ($Folder in $FolderList) {
         }
 
         catch {
-            Write-Warning "$Folder $EicarFileName file couldn't be created. Either permissions or AV prevented file creation."
+            Write-Warning "$Folder $EicarFullFileName file couldn't be created. Either permissions or AV prevented file creation."
         }
     }
 
     else {
-        Write-SimpleLogfile -string ("[WARNING] - $EicarFileName already exists!: " + $FilePath) -name $LogFile -OutHost
+        Write-SimpleLogfile -string ("[WARNING] - $EicarFullFileName already exists!: " + $FilePath) -name $LogFile -OutHost
     }
 }
 
-# Try to open each EICAR file to force detection
-foreach ($Folder in $FolderList) {
-    $FilePath = (Join-Path $Folder $EicarFileName)
-    if (Test-Path $FilePath -PathType Leaf) {
-        Start-Process $FilePath -ErrorAction SilentlyContinue -WindowStyle Minimized
+# Create a random folder in root path
+$randomString = -join ((65..90) + (97..122) | Get-Random -Count 10 | ForEach-Object { [char]$_ })
+$randomFolder = New-Item -Path (Join-Path (Join-Path $env:SystemDrive '\') "TestExchAVExclusions-$randomString") -ItemType Directory
+$extensionsList = New-Object Collections.Generic.List[string]
+$extensionsList = Get-ExchAVExclusionsExtensions -ExchangePath $ExchangePath -MsiProductMinor ([byte]$serverExchangeInstallDirectory.MsiProductMinor)
+
+if ($randomFolder) {
+    foreach ($extension in $extensionsList) {
+        $filepath = Join-Path $randomFolder "$EicarFileName.$extension"
+        Write-SimpleLogfile -String ("Creating $EicarFileName.$extension file " + $FilePath) -name $LogFile
+
+        if (!(Test-Path -Path $FilePath)) {
+
+            # Try writing the encoded string to a the file
+            try {
+                [byte[]] $EicarBytes = [System.Convert]::FromBase64String($EncodedEicar)
+                [string] $Eicar = [System.Text.Encoding]::UTF8.GetString($EicarBytes)
+                [IO.File]::WriteAllText($FilePath, $Eicar)
+            } catch {
+                Write-Warning "$randomFolder $EicarFileName.$extension file couldn't be created. Either permissions or AV prevented file creation."
+            }
+        } else {
+            Write-SimpleLogfile -string ("[WARNING] - $randomFolder $EicarFileName.$extension  already exists!: ") -name $LogFile -OutHost
+        }
     }
+} else {
+    Write-Warning "We cannot create a folder in root path to test extension exclusions."
 }
+
+Write-SimpleLogfile -String "EICAR Files Created" -name $LogFile -OutHost
+
+Write-SimpleLogfile -String "Accessing EICAR Files" -name $LogFile -OutHost
+# Try to open each EICAR file to force detection in paths
+$i = 0
+foreach ($Folder in $FolderList) {
+    $FilePath = (Join-Path $Folder $EicarFullFileName)
+    if (Test-Path $FilePath -PathType Leaf) {
+        Write-SimpleLogfile -String ("Opening $EicarFullFileName file " + $FilePath) -name $LogFile
+        Start-Process -FilePath more -ArgumentList """$FilePath""" -ErrorAction SilentlyContinue -WindowStyle Hidden | Out-Null
+    }
+    $i++
+}
+
+# Try to open extensions:
+$i = 0
+foreach ($extension in $extensionsList) {
+    $FilePath = Join-Path $randomFolder "$EicarFileName.$extension"
+    if (Test-Path $FilePath -PathType Leaf) {
+        Write-SimpleLogfile -String ("Opening $EicarFileName.$extension file " + $FilePath) -name $LogFile
+        Start-Process -FilePath more -ArgumentList """$FilePath""" -ErrorAction SilentlyContinue -WindowStyle Hidden | Out-Null
+    }
+    $i++
+}
+
+Write-SimpleLogfile -String "Access EICAR Files Finished" -name $LogFile -OutHost
 
 # Sleeping 5 minutes for AV to "find" the files
-Start-SleepWithProgress -sleeptime 500 -message "Allowing time for AV to Scan"
+Start-SleepWithProgress -sleeptime 300 -message "Allowing time for AV to Scan"
 
 # Create a list of folders that are probably being scanned by AV
 $BadFolderList = New-Object Collections.Generic.List[string]
@@ -187,27 +240,68 @@ Write-SimpleLogfile -string "Testing for EICAR files" -name $LogFile -OutHost
 # Test each location for the EICAR file
 foreach ($Folder in $FolderList) {
 
-    $FilePath = (Join-Path $Folder $EicarFileName)
+    $FilePath = (Join-Path $Folder $EicarFullFileName)
 
     # If the file exists delete it -- this means the folder is not being scanned
     if (Test-Path $FilePath ) {
-        Write-SimpleLogfile -String ("Removing " + $FilePath) -name $LogFile
-        Remove-Item $FilePath -Confirm:$false -Force
+        #Get content to confirm that the file is not blocked by AV
+        $output = Get-Content $FilePath -ErrorAction SilentlyContinue
+        if ($output -eq $Eicar) {
+            Write-SimpleLogfile -String ("Removing " + $FilePath) -name $LogFile
+            Remove-Item $FilePath -Confirm:$false -Force
+        } else {
+            Write-SimpleLogfile -String ("[FAIL] - Possible AV Scanning on Path: " + $Folder) -name $LogFile -OutHost
+            $BadFolderList.Add($Folder)
+        }
     }
     # If the file doesn't exist Add that to the bad folder list -- means the folder is being scanned
     else {
-        Write-SimpleLogfile -String ("[FAIL] - Possible AV Scanning: " + $Folder) -name $LogFile -OutHost
+        Write-SimpleLogfile -String ("[FAIL] - Possible AV Scanning on Path: " + $Folder) -name $LogFile -OutHost
         $BadFolderList.Add($Folder)
     }
 }
 
+$BadExtensionList = New-Object Collections.Generic.List[string]
+# Test each extension for the EICAR file
+foreach ($extension in $extensionsList) {
+
+    $filepath = Join-Path $randomFolder "$EicarFileName.$extension"
+
+    # If the file exists delete it -- this means the extension is not being scanned
+    if (Test-Path $filepath ) {
+        #Get content to confirm that the file is not blocked by AV
+        $output = Get-Content $FilePath -ErrorAction SilentlyContinue
+        if ($output -eq $Eicar) {
+            Write-SimpleLogfile -String ("Removing " + $FilePath) -name $LogFile
+            Remove-Item $FilePath -Confirm:$false -Force
+        } else {
+            Write-SimpleLogfile -String ("[FAIL] - Possible AV Scanning on Extension: " + $extension) -name $LogFile -OutHost
+            $BadExtensionList.Add($extension)
+        }
+    }
+    # If the file doesn't exist Add that to the bad extension list -- means the extension is being scanned
+    else {
+        Write-SimpleLogfile -String ("[FAIL] - Possible AV Scanning on Extension: " + $extension) -name $LogFile -OutHost
+        $BadExtensionList.Add($extension)
+    }
+}
+
+#Delete randon Folder
+Remove-Item $randomFolder
+
 # Report what we found
-if ($BadFolderList.count -gt 0) {
-    $OutputPath = Join-Path $env:LOCALAPPDATA BadFolders.txt
+if ($BadFolderList.count -gt 0 -or $BadExtensionList.Count -gt 0 ) {
+    $OutputPath = Join-Path $env:LOCALAPPDATA BadExclusions.txt
     $BadFolderList | Out-File $OutputPath
+    $BadExtensionList | Out-File $OutputPath -Append
 
     Write-SimpleLogfile -String "Possbile AV Scanning found" -name $LogFile
-    Write-Warning ("Found " + $BadFolderList.count + " folders that are possibly being scanned!")
+    if ($BadFolderList.count -gt 0 ) {
+        Write-Warning ("Found $($BadFolderList.count) of $($FolderList.Count) folders that are possibly being scanned! ")
+    }
+    if ($BadExtensionList.count -gt 0 ) {
+        Write-Warning ("Found $($BadExtensionList.count) of $($extensionsList.Count) extensions that are possibly being scanned! ")
+    }
     Write-Warning ("Review " + $OutputPath + " For the full list.")
 } else {
     Write-SimpleLogfile -String "All EICAR files found; Exclusions appear to be set properly" -Name $LogFile -OutHost
